Split error definitions into separate file.
